test/requirements_testing.txt
pytest-cov
mock
cassandra-driver==3.25.0
pandas
kubernetes